Benchmarks inlined with modules plus build mods.

+++ circuits/cpp/barretenberg/src/aztec/plonk_bench/plonk.bench.cpp
@@ -1,16 +1,10 @@
 #include <benchmark/benchmark.h>
-
-#include <math.h>
-
-#include <barretenberg/curves/bn254/fr.hpp>
-
-#include <barretenberg/numeric/bitop/get_msb.hpp>
-#include <barretenberg/waffle/composer/standard_composer.hpp>
-#include <barretenberg/waffle/proof_system/preprocess.hpp>
-#include <barretenberg/waffle/proof_system/prover/prover.hpp>
-#include <barretenberg/waffle/proof_system/verifier/verifier.hpp>
-
-#include <barretenberg/waffle/stdlib/field/field.hpp>
+#include <ecc/curves/bn254/fr.hpp>
+#include <numeric/bitop/get_msb.hpp>
+#include <plonk/composer/standard_composer.hpp>
+#include <plonk/proof_system/prover/prover.hpp>
+#include <plonk/proof_system/verifier/verifier.hpp>
+#include <stdlib/primitives/field/field.hpp>
 
 using namespace benchmark;
 
@@ -54,7 +48,7 @@ void construct_proving_keys_bench(State& state) noexcept
         waffle::StandardComposer composer = waffle::StandardComposer(static_cast<size_t>(state.range(0)));
         generate_test_plonk_circuit(composer, static_cast<size_t>(state.range(0)));
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         composer.compute_proving_key();
         state.PauseTiming();
         provers[idx] = composer.preprocess();
@@ -70,7 +64,7 @@ void construct_instances_bench(State& state) noexcept
         waffle::StandardComposer composer = waffle::StandardComposer(static_cast<size_t>(state.range(0)));
         generate_test_plonk_circuit(composer, static_cast<size_t>(state.range(0)));
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         composer.preprocess();
         state.ResumeTiming();
         verifiers[idx] = composer.create_verifier();
@@ -82,7 +76,7 @@ void construct_proofs_bench(State& state) noexcept
 {
     for (auto _ : state) {
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         // provers[idx].reset();
         proofs[idx] = provers[idx].construct_proof();
         state.PauseTiming();
@@ -96,7 +90,7 @@ void verify_proofs_bench(State& state) noexcept
 {
     for (auto _ : state) {
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         verifiers[idx].verify_proof(proofs[idx]);
         state.PauseTiming();
         // if (!result)
@@ -112,7 +106,7 @@ void compute_wire_coefficients(State& state) noexcept
 {
     for (auto _ : state) {
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         provers[idx].reset();
         provers[idx].init_quotient_polynomials();
         provers[idx].compute_wire_coefficients();
@@ -124,7 +118,7 @@ void compute_wire_commitments(State& state) noexcept
 {
     for (auto _ : state) {
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         provers[idx].reset();
         provers[idx].compute_wire_commitments();
     }
@@ -135,7 +129,7 @@ void compute_z_coefficients(State& state) noexcept
 {
     for (auto _ : state) {
         size_t idx =
-            static_cast<size_t>(numeric::get_msb(state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
+            static_cast<size_t>(numeric::get_msb((uint64_t)state.range(0))) - static_cast<size_t>(numeric::get_msb(START));
         provers[idx].compute_z_coefficients();
     }
 }
